requirements.txt
Django<5.0
djangorestframework>=3.12
djangorestframework-simplejwt~=5.3.1
django-cors-headers
channels==3.0.5
channels-redis
django-celery-beat
celery~=5.4.0
PyJWT~=2.10.1

pillow

django-bootstrap-v5
python-dotenv~=1.0.1
selenium~=4.27.1
daphne

icecream
colorlog

redis~=5.2.1
asgiref~=3.8.1
